+++ changelog.md
@@ -9,6 +9,14 @@ All notable changes to HVCWISE_TEA will be documented in this file.
 ### Removed
 ### Fixed
 
+## [matlab_compatibility] - 2025-04-17
+### Added
+- Production per type in KPI
+### Changed
+- Octave sripts compatible with Matlab
+- Correlation > 0 be allowed (with Matlab, not with Octave) for contingencies in the poles of a DC component
+- User guide in the ReadMe
+
 ## [add_tests] - 2025-02-07
 ### Added
 - Number of parallel lines in AC branches